Fix lazy init from config for sources
The global source was initialized too early and therefore did
not support configuration via App.config.

Also, not invoking the .Listeners property, caused the Init to
happen lazily too, which caused the configuration to potentially
be read for the wrong source, since we replace the source name
at tracing time.