Fixed TracerManager to keep TraceSource instances alive (#6)
In some cases, the weak references associated with the trace sources
were been pruned after being configured (added listener, set level, etc).

So, the next traces were ignored because the trace source was created again,
but now without any associated listener or level.

We ensure that the weak reference are not being pruned by keeping a dictionary
of source we track. Additionally the GetOrAdd was improved by looking in the dictionary first.